+++ examples/src/test/scala_2/bigdatatypes/CassandraToOthers.scala
@@ -21,7 +21,6 @@ class CassandraToOthers extends UnitSpec {
       .withColumn("bar", DataTypes.INT)
 
   "Cassandra table" should "be converted into Spark Schema" in {
-    // val sparkSchema: StructType = myDataFrame.schema
     val sparkSchema: StructType = StructType(
       List(
         StructField("id", StringType, nullable = false),